Revert decision on place-content; place-items should be all we need
The problem with place-content is that it squishes it's content. That’s not what we want. All we want is that ZStack childre are centered with respect to the parent bounding box. We don’t also want the parent bounding box to collapse on itself. That being said, we may want to add align-self: stretch? That much isn’t clear yet.